computeTouch

+++ ComputeTouch/src/computeTouchMeasurements/Point.java
@@ -4,14 +4,14 @@ public class Point {
     
     public int X;    
     public int Y;    
-    public double T;
+    public int T;
     
     public Point(int x, int y) {
     	X = x;
     	Y = y;
     }
     
-    public Point(int x, int y, double t) {
+    public Point(int x, int y, int t) {
     	X = x;
     	Y = y;
     	T = t;
